Safari-609.1.11.1/LayoutTests/fast/dom/Element/scrolling-funtions-on-element-expected.txt

Tests scrollTo/By(x,y) and scrollTo/By(scrollOptions) on Element

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS scroller.scrollLeft is 20
PASS scroller.scrollTop is 500
PASS scroller.scrollLeft is 70
PASS scroller.scrollTop is 560
PASS scroller.scrollLeft is 100
PASS scroller.scrollTop is 400
PASS scroller.scrollLeft is 120
PASS scroller.scrollTop is 430
PASS scroller.scrollLeft is 120
PASS scroller.scrollTop is 700
PASS scroller.scrollLeft is 300
PASS scroller.scrollTop is 700
PASS scroller.scrollLeft is 500
PASS scroller.scrollTop is 700
PASS scroller.scrollLeft is 500
PASS scroller.scrollTop is 950
PASS scroller.scrollLeft is 20
PASS scroller.scrollTop is 25
PASS scroller.scrollLeft is 30
PASS scroller.scrollTop is 35
PASS scroller.scrollLeft is 0
PASS scroller.scrollTop is 0
PASS scroller.scrollLeft is 10
PASS scroller.scrollTop is 10
PASS scroller.scrollLeft is 0
PASS scroller.scrollTop is 0
PASS scroller.scrollLeft is 0
PASS scroller.scrollTop is 0
PASS scroller.scrollLeft is 0
PASS scroller.scrollTop is 0
PASS scroller.scrollLeft is 100
PASS scroller.scrollTop is 400
PASS successfullyParsed is true

TEST COMPLETE
